fix: t parameter in hit

+++ src/renderer.jl
@@ -15,7 +15,7 @@ end
 
 struct Scene
     camera::Camera
-    screensize::Int
+    screensize::Float64
     meshes::Vector{Mesh}
 end
